added view details

+++ src/components/pages/Mylist.jsx
@@ -2,6 +2,7 @@
 import { useEffect, useState } from "react";
 import useAuth from "../../Hook/useAuth";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 const Mylist = () => {
@@ -17,16 +18,39 @@ fetch( `http://localhost:5000/mylist/${user?.email}`)
             setMyItem(data);
         });
 },[user,control]);
+                          // DElete :-->
 const handleDelete =(id)=>{
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!"
+  }).then((result)=>{
+if(result.isConfirmed){
   fetch(`http://localhost:5000/delete/${id}`,{
     method:"DELETE",
   })
 .then((res)=>res.json())
  .then((data)=>{
     if(data.deletedCount > 0){
+      Swal.fire(
+         "Deleted!",
+         "Your file has been deleted.",
+         "success"
+      );
       setControl(!control)
     }
 });
+
+}
+
+  })
+
+
+  
 
 }
 
@@ -41,11 +65,11 @@ const handleDelete =(id)=>{
 {item.email}
         <div className="badge badge-secondary">item</div>
       </h2>
-      <p>If a dog chews shoes whose shoes does he choose?</p>
+      <p>{item.short_description}</p>
       <div className="card-actions justify-end">
         <Link to={`/update/${item._id}`}><button className="badge btn badge-outline">Update</button> </Link>
-        <button onClick={()=>handleDelete(item._id)} className="badge btn badge-outline">Delete</button>
-      </div>
+        <button onClick={()=>handleDelete(item._id)} className="badge btn btn badge-outline">Delete</button>
+      </div>   
     </div>
   </div> </div>  )
